Add authorization to favorites.

+++ app/controllers/favorites_controller.rb
@@ -3,7 +3,7 @@ class FavoritesController < ApplicationController
   def create
     @post = Post.find(params[:post_id])
     favorite = current_user.favorites.build(post: @post)
-  
+    authorize favorite
     if favorite.save
       flash[:notice] = "Post was favorited."
       redirect_to [@post.topic, @post]
@@ -17,7 +17,7 @@ class FavoritesController < ApplicationController
   def destroy
     @post = Post.find(params[:post_id])
     favorite = current_user.favorites.find(params[:id])
-
+    authorize favorite
     if favorite.destroy
       flash[:notice] = "Unfavorited successfully."
       redirect_to [@post.topic, @post]
